Improve webhooks page
