+++ apps/web/app/app.dub.co/(dashboard)/[slug]/settings/webhooks/page.tsx
@@ -1,5 +1,5 @@
 import WebhooksPageClient from "./page-client";
 
-export default async function WebhooksPage() {
+export default function WebhooksPage() {
   return <WebhooksPageClient />;
 }
